+++ Video.cpp
@@ -2,6 +2,7 @@
 
 
 #include <assert.h>
+#include <cstring>
 
 #include "Cartridge.h"
 
@@ -42,8 +43,8 @@ void Video::MapCartridgeCHRToPPU()
     /* TODO(Jonathan): Support more mappers here */
     const Cartridge::Header &header = cartridge->GetHeader();
     const byte * const cartridgeRom = cartridge->GetRom();
-    const u32 offset = (header.prgRomSizeKB * 1_KB) + 0x0010;
-    const u32 dataSize = header.chrRomSizeKB * 1_KB;
+    const u64 offset = (header.prgRomSizeKB * 1_KB) + 0x0010;
+    const u64 dataSize = header.chrRomSizeKB * 1_KB;
     memcpy( memory, &cartridgeRom[ offset ], dataSize );
 }
